restructure of codebase
