+++ internal/plugins/install_test.go
@@ -4,12 +4,14 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
+
+	"github.com/sphireinc/foundry/internal/installutil"
 )
 
 func TestSafeArchivePathRejectsTraversal(t *testing.T) {
 	root := t.TempDir()
 
-	if _, err := safeArchivePath(root, "../escape.txt"); err == nil {
+	if _, err := installutil.SafeArchivePath(root, "../escape.txt"); err == nil {
 		t.Fatal("expected path traversal entry to be rejected")
 	}
 }
@@ -17,7 +19,7 @@ func TestSafeArchivePathRejectsTraversal(t *testing.T) {
 func TestSafeArchivePathAcceptsNestedPath(t *testing.T) {
 	root := t.TempDir()
 
-	path, err := safeArchivePath(root, "plugin/assets/app.css")
+	path, err := installutil.SafeArchivePath(root, "plugin/assets/app.css")
 	if err != nil {
 		t.Fatalf("expected nested path to be accepted: %v", err)
 	}
